update seeder encore

+++ src/main/java/matser2/istic/mmmback/DTO/WorksiteAllDto.java
@@ -14,8 +14,8 @@ public class WorksiteAllDto {
     private Date statusUpdated;
     private int duration;
     private Date createdAt;
-    private int longitude;
-    private int latitude;
+    private double longitude;
+    private double latitude;
     private String location;
     private WorksiteStatus status;
     private CustomerGetDto customer;
@@ -122,19 +122,19 @@ public class WorksiteAllDto {
         this.createdAt = createdAt;
     }
 
-    public int getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public int getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(int latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
